Update init.sls, pillar and minio.etc.default
**init.sls:**
  - Add missing '%' in the variables declaration.
  - Add missing ':' in the `mode` of:
      - minio_systemd_service
      - minio_etc_default
  - Add missing whitespace before '}}' in minio_etc_default.

**minio.etc.default.jinja:**
  - Update MINIO_BROWSER, default value should be `off`
  - Update MINIO_DRIVE_SYNC, default value should be `on`

**pillar.example:**
  - Add NOTE regarding MINIO_BROWSER and MINIO_DRIVE_SYNC

browser and drive_sync options (pillar.example) will be ignored
They are keep here for informational purposes.
Sadly "on" and "off" are intepreted as booleans by yaml
so "True" and "False" are the values that are rendered
in the /etc/default/minio file, and obviously minio just
refuses to start as it doesn't recognises that values.