Update install, systemd, user, pillar, templates.
**install.sls:**
  - Add minio_user variable
  - Add minio_group variable
  - Add etc_minio file.directory

The etc_minio file.directory should create the `/etc/minio/certs`
and set the permissions for the minio user and group. Even if minio
is configured to run without SSL, it will refuse to start if it can't
access the certs directory.

**pillar.example:**
  - Update *http_trace* set a defalt log file.
  - Updte replace *opts* with *port* option.

Minio disables the logging by default(http_trace: set to blank), but
yaml renders `None`, and the server refuses to start in this condition.
Set a default logfile to avoid that.
Replace *opts* with *port*. Eventually we can add *address* but with
the issues regarding blank options, better to not add it right now.

**minio.etc.default.jinja:**
  - Replace MINIO_OPTS with MINIO_PORT

**minio.service.jinja:**
  - Update `ExecStart`, replace MINIO_OPTS with MINIO_PORT

**systemd.sls:**
  - Replace `minio_opts` with `minio_port`
  - Add minio_enable_service
  - Add minio_start_service

**user.sls:**
  - Update group task ID
  - Delete group.present.system
  - Update user task ID
  - Add user.present.name